prevent SessionCreateFactory.options from being None
SAMI2PStreamClientEndpoint.new and endpoints.SAMI2PStreamServerEndpoint.new
previously passed their default value of `None` for options onwards to
`getSession` which passed it on to `SessionCreateFactory` where it override
the default value of `{}` and caused `SessionCreateFactory.options` to be
`None`.  This case was not handled in `SessionCreateSender` causing an attempt
to iterate over `None` in `sendSessionCreate`, raising an exception and
preventing session creation.

Now `SessionCreateFactory` makes the default options `None` and takes care to
convert this to an empty dictionary during `__init__`.